+++ packages/framer-motion/src/projection/node/create-projection-node.ts
@@ -86,8 +86,6 @@ const metrics = {
 
 const transformAxes = ["", "X", "Y", "Z"]
 
-const hiddenVisibility: ResolvedValues = { visibility: "hidden" }
-
 /**
  * We use 1000 as the animation target as 0-1000 maps better to pixels than 0-1
  * which has a noticeable difference in spring animations
@@ -461,11 +459,22 @@ export function createProjectionNode<I>({
 
             if (attachResizeListener) {
                 let cancelDelay: VoidFunction
+                let innerWidth = 0
 
                 const resizeUnblockUpdate = () =>
                     (this.root.updateBlockedByResize = false)
 
+                // Set initial innerWidth in a frame.read callback to batch the read
+                frame.read(() => {
+                    innerWidth = window.innerWidth
+                })
+
                 attachResizeListener(instance, () => {
+                    const newInnerWidth = window.innerWidth
+                    if (newInnerWidth === innerWidth) return
+
+                    innerWidth = newInnerWidth
+
                     this.root.updateBlockedByResize = true
 
                     cancelDelay && cancelDelay()
@@ -645,6 +654,7 @@ export function createProjectionNode<I>({
 
         willUpdate(shouldNotifyListeners = true) {
             this.root.hasTreeAnimated = true
+
             if (this.root.isUpdateBlocked()) {
                 this.options.onExitComplete && this.options.onExitComplete()
                 return
@@ -723,30 +733,31 @@ export function createProjectionNode<I>({
                 return
             }
 
-            if (!this.isUpdating) {
-                this.nodes!.forEach(clearIsLayoutDirty)
-            }
-
             this.animationCommitId = this.animationId
 
-            this.isUpdating = false
+            if (!this.isUpdating) {
+                this.nodes!.forEach(clearIsLayoutDirty)
+            } else {
+                this.isUpdating = false
 
-            /**
-             * Write
-             */
-            this.nodes!.forEach(resetTransformStyle)
+                /**
+                 * Write
+                 */
+                this.nodes!.forEach(resetTransformStyle)
 
-            /**
-             * Read ==================
-             */
-            // Update layout measurements of updated children
-            this.nodes!.forEach(updateLayout)
+                /**
+                 * Read ==================
+                 */
+                // Update layout measurements of updated children
+                this.nodes!.forEach(updateLayout)
 
-            /**
-             * Write
-             */
-            // Notify listeners that the layout is updated
-            this.nodes!.forEach(notifyLayoutUpdate)
+                /**
+                 * Write
+                 */
+                // Notify listeners that the layout is updated
+                this.nodes!.forEach(notifyLayoutUpdate)
+            }
+
             this.clearAllSnapshots()
 
             /**
@@ -857,7 +868,6 @@ export function createProjectionNode<I>({
         updateLayout() {
             if (!this.instance) return
 
-            // TODO: Incorporate into a forwarded scroll offset
             this.updateScroll()
 
             if (
@@ -1887,15 +1897,15 @@ export function createProjectionNode<I>({
             visualElement.scheduleRender()
         }
 
-        getProjectionStyles(styleProp?: MotionStyle) {
-            if (!this.instance || this.isSVG) return undefined
+        applyProjectionStyles(
+            targetStyle: any, // CSSStyleDeclaration - doesn't allow numbers to be assigned to properties
+            styleProp?: MotionStyle
+        ) {
+            if (!this.instance || this.isSVG) return
 
             if (!this.isVisible) {
-                return hiddenVisibility
-            }
-
-            const styles: ResolvedValues = {
-                visibility: "",
+                targetStyle.visibility = "hidden"
+                return
             }
 
             const transformTemplate = this.getTransformTemplate()
@@ -1903,61 +1913,64 @@ export function createProjectionNode<I>({
             if (this.needsReset) {
                 this.needsReset = false
 
-                styles.opacity = ""
-                styles.pointerEvents =
+                targetStyle.visibility = ""
+                targetStyle.opacity = ""
+                targetStyle.pointerEvents =
                     resolveMotionValue(styleProp?.pointerEvents) || ""
-                styles.transform = transformTemplate
+                targetStyle.transform = transformTemplate
                     ? transformTemplate(this.latestValues, "")
                     : "none"
-                return styles
+                return
             }
 
             const lead = this.getLead()
             if (!this.projectionDelta || !this.layout || !lead.target) {
-                const emptyStyles: ResolvedValues = {}
                 if (this.options.layoutId) {
-                    emptyStyles.opacity =
+                    targetStyle.opacity =
                         this.latestValues.opacity !== undefined
                             ? this.latestValues.opacity
                             : 1
-                    emptyStyles.pointerEvents =
+                    targetStyle.pointerEvents =
                         resolveMotionValue(styleProp?.pointerEvents) || ""
                 }
                 if (this.hasProjected && !hasTransform(this.latestValues)) {
-                    emptyStyles.transform = transformTemplate
+                    targetStyle.transform = transformTemplate
                         ? transformTemplate({}, "")
                         : "none"
                     this.hasProjected = false
                 }
 
-                return emptyStyles
+                return
             }
+
+            targetStyle.visibility = ""
 
             const valuesToRender = lead.animationValues || lead.latestValues
             this.applyTransformsToTarget()
 
-            styles.transform = buildProjectionTransform(
+            let transform = buildProjectionTransform(
                 this.projectionDeltaWithTransform!,
                 this.treeScale,
                 valuesToRender
             )
 
             if (transformTemplate) {
-                styles.transform = transformTemplate(
-                    valuesToRender,
-                    styles.transform
-                )
+                transform = transformTemplate(valuesToRender, transform)
             }
 
+            targetStyle.transform = transform
+
             const { x, y } = this.projectionDelta
-            styles.transformOrigin = `${x.origin * 100}% ${y.origin * 100}% 0`
+            targetStyle.transformOrigin = `${x.origin * 100}% ${
+                y.origin * 100
+            }% 0`
 
             if (lead.animationValues) {
                 /**
                  * If the lead component is animating, assign this either the entering/leaving
                  * opacity
                  */
-                styles.opacity =
+                targetStyle.opacity =
                     lead === this
                         ? valuesToRender.opacity ??
                           this.latestValues.opacity ??
@@ -1970,7 +1983,7 @@ export function createProjectionNode<I>({
                  * Or we're not animating at all, set the lead component to its layout
                  * opacity and other components to hidden.
                  */
-                styles.opacity =
+                targetStyle.opacity =
                     lead === this
                         ? valuesToRender.opacity !== undefined
                             ? valuesToRender.opacity
@@ -1995,14 +2008,14 @@ export function createProjectionNode<I>({
                  * a corresponding layout animation.
                  */
                 const corrected =
-                    styles.transform === "none"
+                    transform === "none"
                         ? valuesToRender[key]
                         : correct(valuesToRender[key], lead)
 
                 if (applyTo) {
                     const num = applyTo.length
                     for (let i = 0; i < num; i++) {
-                        styles[applyTo[i]] = corrected
+                        targetStyle[applyTo[i] as any] = corrected
                     }
                 } else {
                     // If this is a CSS variable, set it directly on the instance.
@@ -2013,7 +2026,7 @@ export function createProjectionNode<I>({
                             this.options.visualElement as HTMLVisualElement
                         ).renderState.vars[key] = corrected
                     } else {
-                        styles[key] = corrected
+                        targetStyle[key as any] = corrected
                     }
                 }
             }
@@ -2024,13 +2037,11 @@ export function createProjectionNode<I>({
              * pointer events on the lead.
              */
             if (this.options.layoutId) {
-                styles.pointerEvents =
+                targetStyle.pointerEvents =
                     lead === this
                         ? resolveMotionValue(styleProp?.pointerEvents) || ""
                         : "none"
             }
-
-            return styles
         }
 
         clearSnapshot() {