+++ CHANGELOG.md
@@ -4,6 +4,32 @@ Motion adheres to [Semantic Versioning](http://semver.org/).
 
 Undocumented APIs should be considered internal and may change without warning.
 
+## [12.20.5] 2025-07-01
+
+### Fixed
+
+-   Improve performance of projection styles by writing directly to `element.style`.
+-   Only block layout animations when window has changed width.
+
+## [12.20.4] 2025-07-01
+
+### Fixed
+
+-   Improve performance of render loop.
+-   Improve performance of types.
+
+## [12.20.3] 2025-07-01
+
+### Fixed
+
+-   Always skip layout animation check on tree mount.
+
+## [12.20.2] 2025-07-01
+
+### Fixed
+
+-   Updated `Transition` type to be more permissive.
+
 ## [12.20.1] 2025-06-30
 
 ### Fixed